Fix problems with comparing fixed sized arrays now that Rust no longer coerces them into slices
Recent changes to Rust modified how the Eq traits works such that
it no longer coerces fixed sized arrays to slices. This results
in issues if one side of the comparison is a slice and the other
is a fixed sized array or if the arrays have more than 32 elements.