Remove rustp-crypto-util
The idea of rust-crypto-util was to have a tool to
compare the outputs of algorithms implemented in
rust-crypto against the same algorithms written in
another language - specifically Java. This was only
ever implemented for Scrypt, however, and has broken
many, many times without have caught any issues. With
the switch to a pure-Cargo build, it becomes a bit harder
to run. This commit removes it since it doesn't appear to
be carrying its weight at the present time. If it could
be useful in the future, it could be re-introduced,
however.